Add allow_failures to .travis.yml matrix
Previously, the failure to compile against a development version of
django would fail the entire build. This makes that a development
version that is allowed to fail without failing the entire build.

This makes it easier to showcase the stability of the project, and
see when pull requests actually break things (compared to everything
always failing).