Change suggested usage of 404CatchMiddleware
I ran into an issue with the Sentry middleware logging 404s that we couldn't reproduce in the frontend. It turned out that one of our middlewares was working as expected by catching certain 404s, but as we followed the Raven docs, Sentry reported a 404 for it nonetheless. 
I posted a slightly more elaborate version of what's going here: https://stackoverflow.com/a/28343606/753738