Use cached_property for getting client in the Django middleware
Since we need to lazy import the client inside the @property, we incur
the overhead of importing every time.

This overhead is minimal, ~0.7usec on my machine, but this is done on
every request, 3 times per request which adds up to ~2.1usec.

As a @cached_property, this gets transformed directly into the __dict__
of the instance, and brings the time down to ~0.04usec per lookup, which
is over an order of magnitude faster after the initial call.