Revert "Fix Django LOGGING example root logger"
This reverts commit 5d863173772cb53edcf95c7ba83060ebeac90c38.

In https://docs.python.org/library/logging.config.html#dictionary-schema-details,
there is in fact a special key named ``root`` in the ``LOGGING`` dict
that specifies the configuration for the root logger.
I think it can also be specified in the ``loggers`` dict, but then its key
should just be the empty string rather than ``root``!