In Django HTTP request data, multiple values for same key are lists
In a Django HttpRequest object, the request data is a QueryDict object,
which is a dictionary-like class that supports multiple values for the
same key. This is used, for instance, with <select multiple> form
elements.

In order to have the request data serialize into JSON correctly when
there are multiple values, it is necessary to use QueryDict.iterlists() /
QueryDict.lists() (python 2 and 3, respectively). Otherwise, the value
passed to Sentry will contain only the last value in the query string.