Purge scheme from dsn templatetag
As is, when capturing an exception on http, the proto relative call results in:

> 307 Internal Redirect
> Location:https://app.getsentry.com/api/....

Which is troublesome because it results in CORS error:

> XMLHttpRequest cannot load .... Origin 'null' is therefore not allowed access.

The event is is still successfully logged. At best, this error is annoying. At worst, it is a red herring when debugging.

This change enforces https and simplifies matters by purging the `scheme` argument.