Default `get_public_dsn()` to https
The API enforces https, the DSN needs to as well.

When using proto-relative on http, the call results in:

> 307 Internal Redirect
> Location:https://app.getsentry.com/api/....

Which is troublesome because it results in CORS error:

> XMLHttpRequest cannot load .... Origin 'null' is therefore not allowed access.

The event is is still successfully logged.

The error is, at best, annoying. At worst, is a red herring.

The `scheme` argument can not be purged without breaking backward
compatibility. Folks already using:

```
{% sentry_public_dsn 'https' %}
```

will see no change. Those using:

```
{% sentry_public_dsn %}
```

Will no longer get a console error.