Ensure that all keys in locals are strings before serializing
Trio uses an ugly hack and puts a value into locals with a
non string key (class LOCALS_KI_PROTECTION_ENABLED).

When trying to serialize that, JsonEncoder crashes.
So we ensure that every key is indeed a string.

Another way to solve this problem would be to pass skipkeys=True
to json.dumps, but this might mask other errors.